pyproject.toml
[project]
name = "fastapi_forge"
version = "0.1.0"
description = "FastAPI Forge"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.1",
    "bcrypt>=5.0.0",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "PyJWT>=2.10.1",
    "loguru>=0.7.3",
    "psutil>=7.2.1",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "scalar-fastapi>=1.6.0",
    "sqlmodel>=0.0.31",
    "orjson>=3.11.5",
]
